refactor(electron): finalize lib directory cleanup
Move ThunkManager.ts from src/lib/ to src/thunk/ThunkManager.ts and remove empty lib directory.

Updated import paths in:
- src/action/ActionScheduler.ts
- src/thunk/init.ts
- src/thunk/registration/ThunkRegistrationQueue.ts
- test/thunk/ThunkManager.spec.ts

All tests continue to pass (469/469) and validation suite clean.
The src/lib/ directory has been completely eliminated.
