+++ packages/electron/src/thunk/ThunkManager.ts
